folly: Bits: make bitsPerBlock 64 bit back
Summary: D1770613 made bitsPerBlock 32 bit. Sadpanda.

Test Plan: _bin/unicorn/diskindex4/test/doc_map_test --gtest_filter=DocMap/DocMapTest.Deleted/23

Reviewed By: meyering@fb.com

Subscribers: net-systems@, folly-diffs@

FB internal diff: D1776343

Tasks: 5968993

Signature: t1:1776343:1420852923:f35166dd9023a701ff16467710131f821810215a

Blame Revision: D1770613

+++ folly/Bits.h
@@ -525,7 +525,7 @@ class BitIterator
       (other.bitOffset_ - bitOffset_);
   }
 
-  unsigned int bitOffset_;
+  size_t bitOffset_;
 };
 
 /**
